% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_add_tasks.R
\name{dag_add_tasks}
\alias{dag_add_tasks}
\title{Add Tasks to a DAG}
\usage{
dag_add_tasks(env, ...)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{...}{Any number of ConnectTask R6 environments created by \link[connectapi.dag]{connect_task}}
}
\description{
Adds ConnectTask environments to an existing ConnectDAG
}
\examples{
task0 <- connect_task("task0", simulated = TRUE)
task1 <- connect_task("task1", simulated = TRUE)
task0 |> set_downstream(task1)

my_dag <-
  connect_dag() |>
  dag_add_tasks(task0, task1)

my_dag
}
