% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_remove_task.R
\name{dag_remove_task}
\alias{dag_remove_task}
\title{Remove a task from a DAG}
\usage{
dag_remove_task(env, task)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{task}{A connectTask R6 environment created by \link[connectapi.dag]{connect_task}}
}
\description{
Removes a specified ConnectTask environment from a ConnectDAG environment.
}
\examples{
task0 <- connect_task("task0", simulated = TRUE)
task1 <- connect_task("task1", simulated = TRUE)

my_dag <-
  connect_dag() |>
  dag_add_tasks(task0, task1)

my_dag |> dag_remove_task(task1)
my_dag$tasks
}
