% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_set_max_concurrent.R
\name{dag_set_dag_timeout}
\alias{dag_set_dag_timeout}
\title{Set the overall timeout for a DAG run}
\usage{
dag_set_dag_timeout(env, seconds)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{seconds}{A positive number of seconds, or NA to disable the timeout}
}
\description{
Sets the maximum number of seconds an entire DAG run may take. When the limit
is exceeded, any still-running tasks are marked Failed and the run stops,
guaranteeing the scheduler terminates. Pass `NA` to disable the global timeout
(the default).
}
\examples{
my_dag <- connect_dag(name = "my_dag")
dag_set_dag_timeout(my_dag, 3600)
}
