% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_set_name.R
\name{dag_set_name}
\alias{dag_set_name}
\title{Set the name of a DAG}
\usage{
dag_set_name(env, name)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{name}{A scalar character of the name you want to set for the DAG}
}
\description{
DAG names are used as identifiers for the DAG when using pins.
}
\examples{
my_dag <-
  connect_dag() |>
  dag_set_name("My Awesome DAG")

print(my_dag)
}
