% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_set_pin_name.R
\name{dag_set_pin_name}
\alias{dag_set_pin_name}
\title{Sets the Pin name for Connect Board}
\usage{
dag_set_pin_name(env, pin_name)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{pin_name}{A scalar character you want to use for the pin name}
}
\description{
When using this package's built-in features for the `pins` package,
the ConnectDAG's pin_name is used to write to a Connect Board.
This function allows you to overwrite the default name given for the Pin Name.
By default the pin name will be the ConnectDAG's name,
but will substitute any invalid pin name characters with a hyphen.
Note that using `dag_set_name()` will also change the pin name.
}
\examples{
my_dag <- connect_dag(name = "my_dag")
my_dag |> dag_set_pin_name("pinned_dag_name")

my_dag$pin_name
}
