% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_set_max_concurrent.R
\name{dag_set_poll_interval}
\alias{dag_set_poll_interval}
\title{Set the scheduler poll interval for a DAG}
\usage{
dag_set_poll_interval(env, seconds)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{seconds}{A positive number of seconds}
}
\description{
Sets how many seconds the scheduler sleeps between poll cycles while tasks are
running. Lower values make the DAG notice finished tasks sooner at the cost of
more frequent requests to Posit Connect. Defaults to 1 second.
}
\examples{
my_dag <- connect_dag(name = "my_dag")
dag_set_poll_interval(my_dag, 5)
}
