% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_write_connect_pin.R
\name{dag_write_connect_pin}
\alias{dag_write_connect_pin}
\title{Save a DAG as a Pin to Posit Connect}
\usage{
dag_write_connect_pin(
  env,
  additional_tags = NA_character_,
  pin_title = paste(env$name, "connectapi.dag"),
  pin_description = NA_character_,
  ...
)
}
\arguments{
\item{env}{A ConnectDAG R6 environment created by \link[connectapi.dag]{connect_dag}}

\item{additional_tags}{A character vector of tags to include on the pin. "DAG" is always applied.}

\item{pin_title}{Passed to pins::pin_write(title)}

\item{pin_description}{Passed to pins::pin_write(description)}

\item{...}{arguments passed to \link[pins]{board_connect}}
}
\description{
Uses the `pins` package to save the DAG to a Connect Board.
Pinned DAGs can be easily loaded and run by scheduled jobs.
After the DAG runs, it can be saved again and versioned.
This allows you to load specific instances of DAG runs, and evaluate how it ran.
}
\examples{
\dontrun{
task0 <- connect_task("task0", simulated = TRUE)
task1 <- connect_task("task1", simulated = TRUE)
task0 |> set_downstream(task1)

my_dag <- connect_dag(name = "my_dag", task0, task1)
dag_write_connect_pin(my_dag)
}

}
