% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_downstream.R
\name{set_downstream}
\alias{set_downstream}
\title{Sets dependent task(s) for a given ConnecTask}
\usage{
set_downstream(env, ...)
}
\arguments{
\item{env}{A ConnectTask environment generated by \link[connectapi.dag]{connect_task}
that you want to set downstream/dependent tasks for.}

\item{...}{Any ConnectTasks to make downstream of `env`}
}
\value{
Does not return anything, it modified to ConnecTask Environment in-place
}
\description{
Tasks downstream from a given Connect Task will execute after the provided ConnectTask.
When a task is linked as an downstream task, the downstream task also creates an upstream link to the provided task.
}
\examples{
task0 <- connect_task("task0", simulated = TRUE)
task1 <- connect_task("task1", simulated = TRUE)
task2 <- connect_task("task2", simulated = TRUE)

task0 |> set_downstream(task1, task2)
plot(task0)
}
