% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_upstream.R
\name{set_upstream}
\alias{set_upstream}
\title{Sets dependency task(s) for a given ConnecTask}
\usage{
set_upstream(object, ...)
}
\arguments{
\item{object}{A ConnectTask environment generated by \link[connectapi.dag]{connect_task}
that you want to set upstream/dependency tasks for.}

\item{...}{Any ConnectTasks to make upstream of `object`}
}
\value{
Does not return anything, it modified to ConnecTask Environment in-place
}
\description{
Tasks upstream from a given Connect Task must complete evaluation before the task runs.
When a task is linked as an upstream task, the upstream task also creates a downstream link to the provided task.
}
\examples{
task0 <- connect_task("task0", simulated = TRUE)
task1 <- connect_task("task1", simulated = TRUE)
task2 <- connect_task("task2", simulated = TRUE)

task2 |> set_upstream(task0, task1)
plot(task2)
}
