% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_run.R
\name{task_run}
\alias{task_run}
\title{Run a single ConnectTask}
\usage{
task_run(env, verbose = FALSE)
}
\arguments{
\item{env}{a ConnectTask R6 environment created by \link[connectapi.dag]{connect_task}}

\item{verbose}{A boolean, when TRUE prints messages to console as the task executes}
}
\description{
Executes a ConnectTask, causing the content on Connect to re-render.
This is not normally called directly, the ConnectDAG will use this function instead.
}
\examples{
task0 <- connect_task("task0", simulated = TRUE)
task_run(task0)
}
