% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_connect_dag.R
\name{validate_connect_dag}
\alias{validate_connect_dag}
\title{Validates a ConnectDAG is complete and can be executed}
\usage{
validate_connect_dag(env)
}
\arguments{
\item{env}{A ConnectDAG R6 environment}
}
\value{
A boolean indicating if it is valid or not
}
\description{
A valid ConnectDAG has all tasks linked in a single network (no islands),
and is a [graph] DAG.
A DAG that is not valid will generate a warning message with the reason.
}
\examples{
task1 <- connect_task("task1", simulated = TRUE)
task2 <- connect_task("task2", simulated = TRUE)
task1 |> set_downstream(task2)

my_dag <- connect_dag(task1, task2)
validate_connect_dag(my_dag)
}
