% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTask.R
\name{sim_task}
\alias{sim_task}
\title{Simulate a Connect Task}
\usage{
sim_task(task_name, trigger_rule = "all_success", fail_prob = 0.2)
}
\arguments{
\item{task_name}{A scalar character for the name of the task deployed to Posit Connect}

\item{trigger_rule}{A scalar character that defines state of dependency (upstream) tasks must be in to execute}

\item{fail_prob}{A value between 0 and 1 for the probability of task failure}
}
\description{
Simulates a connect task evaluation, with an optional probability for the task to fail.
}
